Removed starting click from TTS
The first click was an artefact of how various things were being put together. Which resulted in the initial portion of the audio file that is played to contain a consistent clicking noise.

Also found that the end click is not the same, but rather an artefact of how the file is played, and the sound it makes when it ends.