Removed clicking sound
TTS is now clear, and there is no clicking noise.

It seems the use of pausing the audio early helped out with getting rid of the clicking at the end.
Setting the values to 0 (silenced) at the end proved ineffective, so I removed it due to the additional computational load.

Extended the amountToRemove due to there still being a slight click at the start.