Implemented a blinking system (also enabled mock HMD)
Implemented an animation system that controls the blinking of a model. Where it chooses from one of five animation (which can easily be expanded)  to play over any other animation.

Also added a Mock head mounted display to the project to better see what a user may see while developing. Without always needing to plug in a HMD.