Implemented basic listening animation system
Implemented a rough version of the system that uses an object to store a boolean to be accessed by all virtual humans within the scene.

They then play animations so as to give off a better sense of them listening.

Currently a rough implementation due to their being a slight issue regarding the animation transition. Currently it is getting stuck in a loop which needs to be fixed.