Added GPT-3 access to the project
Added the needed dependencies to access GPT-3 from Unity (the DLLs in the plugins folder).

Also in GPT3.cs we have developed the functions that we believe will be sufficient for the purpose of our specified text variation.

Finally, started shifting away from the api keys being stored with the repository and publicly. To now storing them in a folder called Keys.
The keys will be stored in text files and only held locally, rather than contributed to the repository.
The reason for this decision is not due to this just being bad practice, but also, some of the services involved are premium (paid for), thus, we would prefer our money being spent where and when we intended.