Implemented text variation into the NPC text-to-speech
Now the original message is passed on to our functions, which are referring to variables relevant to the subject that is talking.

We have been working on refining the prompt which we are using to try and get the best response from the model. Which has shown great promise, but still needs to further tweaks.

Changed the yarn script from the basic demo to excerpts from the original demonstration, with one from each NPC.